+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Changed:
 
 - Surface and PictureRecorder cache returned Canvas object and invalidate it when owner is closed #66
+- Fixed ColorType enum mapping #73
 
 # 0.116.2 - Dec 8, 2023